@@ -1,6 +1,7 @@
 {
     "name": "pelican-dev/panel",
     "description": "The free, open-source game management panel. Supporting Minecraft, Spigot, BungeeCord, and SRCDS servers.",
+    "license": "AGPL-3.0-only",
     "require": {
         "php": "^8.2 || ^8.3 || ^8.4",
         "ext-intl": "*",
@@ -9,17 +10,18 @@
         "ext-pdo": "*",
         "ext-zip": "*",
         "abdelhamiderrahmouni/filament-monaco-editor": "^0.2.5",
-        "aws/aws-sdk-php": "^3.342",
+        "aws/aws-sdk-php": "^3.343",
+        "aymanalhattami/filament-context-menu": "^1.0",
         "calebporzio/sushi": "^2.5",
         "chillerlan/php-qrcode": "^5.0.2",
         "dedoc/scramble": "^0.12.10",
         "doctrine/dbal": "~3.6.0",
         "filament/filament": "^3.3",
         "guzzlehttp/guzzle": "^7.9",
-        "laravel/framework": "^12.10",
+        "laravel/framework": "^12.17",
         "laravel/helpers": "^1.7",
-        "laravel/sanctum": "^4.0.2",
-        "laravel/socialite": "^5.19",
+        "laravel/sanctum": "^4.1",
+        "laravel/socialite": "^5.21",
         "laravel/tinker": "^2.10.1",
         "laravel/ui": "^4.6",
         "lcobucci/jwt": "~4.3.0",
@@ -33,10 +35,10 @@
         "socialiteproviders/authentik": "^5.2",
         "socialiteproviders/discord": "^4.2",
         "socialiteproviders/steam": "^4.3",
-        "spatie/laravel-data": "^4.14",
+        "spatie/laravel-data": "^4.15",
         "spatie/laravel-fractal": "^6.3",
         "spatie/laravel-health": "^1.34",
-        "spatie/laravel-permission": "^6.16",
+        "spatie/laravel-permission": "^6.19",
         "spatie/laravel-query-builder": "^6.3",
         "spatie/temporary-directory": "^2.3",
         "symfony/http-client": "^7.2",
@@ -49,13 +51,15 @@
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^3.5",
         "fakerphp/faker": "^1.23.1",
-        "larastan/larastan": "3.x-dev#5bd1c40edb43a727584081e74e9a1a2a201ea2ee",
+        "larastan/larastan": "^3.4",
         "laravel/pail": "^1.2.2",
         "laravel/pint": "^1.15.3",
         "laravel/sail": "^1.41",
         "mockery/mockery": "^1.6.11",
         "nunomaduro/collision": "^8.6",
         "pestphp/pest": "^3.7",
+        "pestphp/pest-plugin-faker": "^3.0",
+        "pestphp/pest-plugin-livewire": "^3.0",
         "spatie/laravel-ignition": "^2.9"
     },
     "autoload": {
