@@ -71,10 +71,10 @@ class DatabasesRelationManager extends RelationManager
             ])
             ->actions([
                 DeleteAction::make()
-                    ->authorize(fn (Database $database) => auth()->user()->can('delete database', $database)),
+                    ->authorize(fn (Database $database) => auth()->user()->can('delete', $database)),
                 ViewAction::make()
                     ->color('primary')
-                    ->hidden(fn () => !auth()->user()->can('viewList database')),
+                    ->hidden(fn () => !auth()->user()->can('viewAny', Database::class)),
             ]);
     }
 }
