@@ -10,6 +10,7 @@ use App\Services\Api\KeyCreationService;
 use Filament\Forms;
 use Filament\Forms\Components\Actions\Action;
 use Filament\Forms\Components\Grid;
+use Filament\Forms\Components\Hidden;
 use Filament\Forms\Components\TagsInput;
 use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\ToggleButtons;
@@ -54,14 +55,14 @@ class CreateNode extends CreateRecord
     {
         $key = ApiKey::query()
             ->where('key_type', ApiKey::TYPE_APPLICATION)
-            ->whereJsonContains('permissions->' . Node::RESOURCE_NAME, AdminAcl::READ|AdminAcl::WRITE)
+            ->whereJsonContains('permissions->' . Node::RESOURCE_NAME, AdminAcl::READ | AdminAcl::WRITE)
             ->first();
 
         if (!$key) {
             $key = $this->keyCreationService->setKeyType(ApiKey::TYPE_APPLICATION)->handle([
                 'memo' => 'Automatically generated node cloud key.',
                 'user_id' => auth()->user()->id,
-                'permissions' => [Node::RESOURCE_NAME => AdminAcl::READ|AdminAcl::WRITE],
+                'permissions' => [Node::RESOURCE_NAME => AdminAcl::READ | AdminAcl::WRITE],
             ]);
         }
 
@@ -83,8 +84,7 @@ class CreateNode extends CreateRecord
                     It looks like there was a problem communicating with Pelican Cloud!
                     Please make a ticket by <a class='underline text-blue-400' href='https://hub.pelican.dev/tickets?code=$code'>clicking here</a>.
                 "))
-                ->send()
-            ;
+                ->send();
 
             return;
         }
@@ -106,8 +106,7 @@ class CreateNode extends CreateRecord
                         ->schema([
                             View::make('filament.admin.nodes.config.left'),
                             View::make('filament.admin.nodes.config.right'),
-                        ])
-                    ,
+                        ]),
                     Step::make('basic')
                         ->label(trans('admin/node.tabs.basic_settings'))
                         ->icon('tabler-server')
@@ -202,15 +201,10 @@ class CreateNode extends CreateRecord
                                     'lg' => 1,
                                 ]),
 
-                            TextInput::make('daemon_listen')
-                                ->columnSpan([
-                                    'default' => 1,
-                                    'sm' => 1,
-                                    'md' => 1,
-                                    'lg' => 1,
-                                ])
-                                ->label(trans('admin/node.port'))
-                                ->helperText(trans('admin/node.port_help'))
+                            TextInput::make('daemon_connect')
+                                ->columnSpan(1)
+                                ->label(fn (Get $get) => $get('connection') === 'https_proxy' ? trans('admin/node.connect_port') : trans('admin/node.port'))
+                                ->helperText(fn (Get $get) => $get('connection') === 'https_proxy' ? trans('admin/node.connect_port_help') : trans('admin/node.port_help'))
                                 ->minValue(1)
                                 ->maxValue(65535)
                                 ->default(8080)
@@ -228,14 +222,15 @@ class CreateNode extends CreateRecord
                                 ->required()
                                 ->maxLength(100),
 
-                            ToggleButtons::make('scheme')
+                            Hidden::make('scheme')
+                                ->default(fn () => request()->isSecure() ? 'https' : 'http'),
+
+                            Hidden::make('behind_proxy')
+                                ->default(false),
+
+                            ToggleButtons::make('connection')
                                 ->label(trans('admin/node.ssl'))
-                                ->columnSpan([
-                                    'default' => 1,
-                                    'sm' => 1,
-                                    'md' => 1,
-                                    'lg' => 1,
-                                ])
+                                ->columnSpan(1)
                                 ->inline()
                                 ->helperText(function (Get $get) {
                                     if (request()->isSecure()) {
@@ -248,20 +243,43 @@ class CreateNode extends CreateRecord
 
                                     return '';
                                 })
-                                ->disableOptionWhen(fn (string $value): bool => $value === 'http' && request()->isSecure())
+                                ->disableOptionWhen(fn (string $value) => $value === 'http' && request()->isSecure())
                                 ->options([
                                     'http' => 'HTTP',
                                     'https' => 'HTTPS (SSL)',
+                                    'https_proxy' => 'HTTPS with (reverse) proxy',
                                 ])
                                 ->colors([
                                     'http' => 'warning',
                                     'https' => 'success',
+                                    'https_proxy' => 'success',
                                 ])
                                 ->icons([
                                     'http' => 'tabler-lock-open-off',
                                     'https' => 'tabler-lock',
+                                    'https_proxy' => 'tabler-shield-lock',
                                 ])
-                                ->default(fn () => request()->isSecure() ? 'https' : 'http'),
+                                ->default(fn () => request()->isSecure() ? 'https' : 'http')
+                                ->live()
+                                ->dehydrated(false)
+                                ->afterStateUpdated(function ($state, Set $set) {
+                                    $set('scheme', $state === 'http' ? 'http' : 'https');
+                                    $set('behind_proxy', $state === 'https_proxy');
+
+                                    $set('daemon_connect', $state === 'https_proxy' ? 443 : 8080);
+                                    $set('daemon_listen', 8080);
+                                }),
+
+                            TextInput::make('daemon_listen')
+                                ->columnSpan(1)
+                                ->label(trans('admin/node.listen_port'))
+                                ->helperText(trans('admin/node.listen_port_help'))
+                                ->minValue(1)
+                                ->maxValue(65535)
+                                ->default(8080)
+                                ->required()
+                                ->integer()
+                                ->visible(fn (Get $get) => $get('connection') === 'https_proxy'),
                         ]),
                     Step::make('advanced')
                         ->label(trans('admin/node.tabs.advanced_settings'))
@@ -485,5 +503,14 @@ class CreateNode extends CreateRecord
     protected function getFormActions(): array
     {
         return [];
+    }
+
+    protected function mutateFormDataBeforeCreate(array $data): array
+    {
+        if (!$data['behind_proxy']) {
+            $data['daemon_listen'] = $data['daemon_connect'];
+        }
+
+        return $data;
     }
 }
