@@ -3,7 +3,6 @@
 namespace App\Filament\Admin\Resources\NodeResource\Widgets;
 
 use App\Models\Node;
-use Carbon\Carbon;
 use Filament\Support\RawJs;
 use Filament\Widgets\ChartWidget;
 use Illuminate\Support\Number;
@@ -16,22 +15,34 @@ class NodeCpuChart extends ChartWidget
 
     public Node $node;
 
+    /**
+     * @var array<int, array{cpu: string, timestamp: string}>
+     */
+    protected array $cpuHistory = [];
+
+    protected int $threads = 0;
+
     protected function getData(): array
     {
-        $threads = $this->node->systemInformation()['cpu_count'] ?? 0;
+        $sessionKey = "node_stats.{$this->node->id}";
 
-        $cpu = collect(cache()->get("nodes.{$this->node->id}.cpu_percent"))
-            ->slice(-10)
-            ->map(fn ($value, $key) => [
-                'cpu' => round($value * $threads, 2),
-                'timestamp' => Carbon::createFromTimestamp($key, auth()->user()->timezone ?? 'UTC')->format('H:i:s'),
-            ])
-            ->all();
+        $data = $this->node->statistics();
+
+        $this->threads = session("{$sessionKey}.threads", $this->node->systemInformation()['cpu_count'] ?? 0);
+
+        $this->cpuHistory = session("{$sessionKey}.cpu_history", []);
+        $this->cpuHistory[] = [
+            'cpu' => round($data['cpu_percent'] * $this->threads, 2),
+            'timestamp' => now(auth()->user()->timezone ?? 'UTC')->format('H:i:s'),
+        ];
+
+        $this->cpuHistory = array_slice($this->cpuHistory, -60);
+        session()->put("{$sessionKey}.cpu_history", $this->cpuHistory);
 
         return [
             'datasets' => [
                 [
-                    'data' => array_column($cpu, 'cpu'),
+                    'data' => array_column($this->cpuHistory, 'cpu'),
                     'backgroundColor' => [
                         'rgba(96, 165, 250, 0.3)',
                     ],
@@ -39,7 +50,7 @@ class NodeCpuChart extends ChartWidget
                     'fill' => true,
                 ],
             ],
-            'labels' => array_column($cpu, 'timestamp'),
+            'labels' => array_column($this->cpuHistory, 'timestamp'),
             'locale' => auth()->user()->language ?? 'en',
         ];
     }
@@ -69,10 +80,10 @@ class NodeCpuChart extends ChartWidget
 
     public function getHeading(): string
     {
-        $threads = $this->node->systemInformation()['cpu_count'] ?? 0;
+        $data = array_slice(end($this->cpuHistory), -60);
 
-        $cpu = Number::format(collect(cache()->get("nodes.{$this->node->id}.cpu_percent"))->last() * $threads, maxPrecision: 2, locale: auth()->user()->language);
-        $max = Number::format($threads * 100, locale: auth()->user()->language);
+        $cpu = Number::format($data['cpu'], maxPrecision: 2, locale: auth()->user()->language);
+        $max = Number::format($this->threads * 100, locale: auth()->user()->language);
 
         return trans('admin/node.cpu_chart', ['cpu' => $cpu, 'max' => $max]);
     }
