@@ -6,5 +6,5 @@ use Filament\Tables\Columns\Column;
 
 class ServerEntryColumn extends Column
 {
-    protected string $view = 'tables.columns.server-entry-column';
+    protected string $view = 'livewire.columns.server-entry-column';
 }
