@@ -19,8 +19,8 @@ use Filament\Forms\Components\Tabs\Tab;
 use Filament\Forms\Components\TextInput;
 use Filament\Forms\Set;
 use Filament\Notifications\Notification;
-use Filament\Tables\Actions\DeleteAction;
 use Filament\Resources\Resource;
+use Filament\Tables\Actions\DeleteAction;
 use Filament\Tables\Actions\EditAction;
 use Filament\Tables\Columns\ImageColumn;
 use Filament\Tables\Columns\TextColumn;
@@ -82,6 +82,35 @@ class UserResource extends Resource
     {
         /** @var Server $server */
         $server = Filament::getTenant();
+
+        $tabs = [];
+        $permissionsArray = [];
+
+        foreach (Permission::permissionData() as $data) {
+            $options = [];
+            $descriptions = [];
+
+            foreach ($data['permissions'] as $permission) {
+                $options[$permission] = str($permission)->headline();
+                $descriptions[$permission] = trans('server/users.permissions.' . $data['name'] . '_' . str($permission)->replace('-', '_'));
+                $permissionsArray[$data['name']][] = $permission;
+            }
+
+            $tabs[] = Tab::make(str($data['name'])->headline())
+                ->schema([
+                    Section::make()
+                        ->description(trans('server/users.permissions.' . $data['name'] . '_desc'))
+                        ->icon($data['icon'])
+                        ->schema([
+                            CheckboxList::make($data['name'])
+                                ->label('')
+                                ->bulkToggleable()
+                                ->columns(2)
+                                ->options($options)
+                                ->descriptions($descriptions),
+                        ]),
+                ]);
+        }
 
         return $table
             ->paginated(false)
@@ -158,69 +187,8 @@ class UserResource extends Resource
                                 Actions::make([
                                     Action::make('assignAll')
                                         ->label('Assign All')
-                                        ->action(function (Set $set) {
-                                            $permissions = [
-                                                'control' => [
-                                                    'console',
-                                                    'start',
-                                                    'stop',
-                                                    'restart',
-                                                ],
-                                                'user' => [
-                                                    'read',
-                                                    'create',
-                                                    'update',
-                                                    'delete',
-                                                ],
-                                                'file' => [
-                                                    'read',
-                                                    'read-content',
-                                                    'create',
-                                                    'update',
-                                                    'delete',
-                                                    'archive',
-                                                    'sftp',
-                                                ],
-                                                'backup' => [
-                                                    'read',
-                                                    'create',
-                                                    'delete',
-                                                    'download',
-                                                    'restore',
-                                                ],
-                                                'allocation' => [
-                                                    'read',
-                                                    'create',
-                                                    'update',
-                                                    'delete',
-                                                ],
-                                                'startup' => [
-                                                    'read',
-                                                    'update',
-                                                    'docker-image',
-                                                ],
-                                                'database' => [
-                                                    'read',
-                                                    'create',
-                                                    'update',
-                                                    'delete',
-                                                    'view_password',
-                                                ],
-                                                'schedule' => [
-                                                    'read',
-                                                    'create',
-                                                    'update',
-                                                    'delete',
-                                                ],
-                                                'settings' => [
-                                                    'rename',
-                                                    'reinstall',
-                                                ],
-                                                'activity' => [
-                                                    'read',
-                                                ],
-                                            ];
-
+                                        ->action(function (Set $set) use ($permissionsArray) {
+                                            $permissions = $permissionsArray;
                                             foreach ($permissions as $key => $value) {
                                                 $allValues = array_unique($value);
                                                 $set($key, $allValues);
@@ -235,264 +203,25 @@ class UserResource extends Resource
                                     ]),
                                 Tabs::make()
                                     ->columnSpanFull()
-                                    ->schema([
-                                        Tab::make('Console')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.control_desc'))
-                                                    ->icon('tabler-terminal-2')
-                                                    ->schema([
-                                                        CheckboxList::make('control')
-                                                            ->formatStateUsing(function (User $user, Set $set) use ($server) {
-                                                                $permissionsArray = $server->subusers->where('user_id', $user->id)->first()->permissions;
-
-                                                                $transformedPermissions = [];
-
-                                                                foreach ($permissionsArray as $permission) {
-                                                                    [$group, $action] = explode('.', $permission, 2);
-                                                                    $transformedPermissions[$group][] = $action;
-                                                                }
-
-                                                                foreach ($transformedPermissions as $key => $value) {
-                                                                    $set($key, $value);
-                                                                }
-
-                                                                return $transformedPermissions['control'] ?? [];
-                                                            })
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'console' => 'Console',
-                                                                'start' => 'Start',
-                                                                'stop' => 'Stop',
-                                                                'restart' => 'Restart',
-                                                            ])
-                                                            ->descriptions([
-                                                                'console' => trans('server/users.permissions.control_console'),
-                                                                'start' => trans('server/users.permissions.control_start'),
-                                                                'stop' => trans('server/users.permissions.control_stop'),
-                                                                'restart' => trans('server/users.permissions.control_restart'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('User')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.user_desc'))
-                                                    ->icon('tabler-users')
-                                                    ->schema([
-                                                        CheckboxList::make('user')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'create' => 'Create',
-                                                                'update' => 'Update',
-                                                                'delete' => 'Delete',
-                                                            ])
-                                                            ->descriptions([
-                                                                'create' => trans('server/users.permissions.user_create'),
-                                                                'read' => trans('server/users.permissions.user_read'),
-                                                                'update' => trans('server/users.permissions.user_update'),
-                                                                'delete' => trans('server/users.permissions.user_delete'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('File')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.file_desc'))
-                                                    ->icon('tabler-folders')
-                                                    ->schema([
-                                                        CheckboxList::make('file')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'read-content' => 'Read Content',
-                                                                'create' => 'Create',
-                                                                'update' => 'Update',
-                                                                'delete' => 'Delete',
-                                                                'archive' => 'Archive',
-                                                                'sftp' => 'SFTP',
-                                                            ])
-                                                            ->descriptions([
-                                                                'create' => trans('server/users.permissions.file_create'),
-                                                                'read' => trans('server/users.permissions.file_read'),
-                                                                'read-content' => trans('server/users.permissions.file_read_content'),
-                                                                'update' => trans('server/users.permissions.file_update'),
-                                                                'delete' => trans('server/users.permissions.file_delete'),
-                                                                'archive' => trans('server/users.permissions.file_archive'),
-                                                                'sftp' => trans('server/users.permissions.file_sftp'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Backup')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.backup_desc'))
-                                                    ->icon('tabler-download')
-                                                    ->schema([
-                                                        CheckboxList::make('backup')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'create' => 'Create',
-                                                                'delete' => 'Delete',
-                                                                'download' => 'Download',
-                                                                'restore' => 'Restore',
-                                                            ])
-                                                            ->descriptions([
-                                                                'create' => trans('server/users.permissions.backup_create'),
-                                                                'read' => trans('server/users.permissions.backup_read'),
-                                                                'delete' => trans('server/users.permissions.backup_delete'),
-                                                                'download' => trans('server/users.permissions.backup_download'),
-                                                                'restore' => trans('server/users.permissions.backup_restore'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Allocation')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.allocation_desc'))
-                                                    ->icon('tabler-network')
-                                                    ->schema([
-                                                        CheckboxList::make('allocation')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'create' => 'Create',
-                                                                'update' => 'Update',
-                                                                'delete' => 'Delete',
-                                                            ])
-                                                            ->descriptions([
-                                                                'read' => trans('server/users.permissions.allocation_read'),
-                                                                'create' => trans('server/users.permissions.allocation_create'),
-                                                                'update' => trans('server/users.permissions.allocation_update'),
-                                                                'delete' => trans('server/users.permissions.allocation_delete'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Startup')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.startup_desc'))
-                                                    ->icon('tabler-question-mark')
-                                                    ->schema([
-                                                        CheckboxList::make('startup')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'update' => 'Update',
-                                                                'docker-image' => 'Docker Image',
-                                                            ])
-                                                            ->descriptions([
-                                                                'read' => trans('server/users.permissions.startup_read'),
-                                                                'update' => trans('server/users.permissions.startup_update'),
-                                                                'docker-image' => trans('server/users.permissions.startup_docker_image'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Database')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.database_desc'))
-                                                    ->icon('tabler-database')
-                                                    ->schema([
-                                                        CheckboxList::make('database')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'create' => 'Create',
-                                                                'update' => 'Update',
-                                                                'delete' => 'Delete',
-                                                                'view_password' => 'View Password',
-                                                            ])
-                                                            ->descriptions([
-                                                                'read' => trans('server/users.permissions.database_read'),
-                                                                'create' => trans('server/users.permissions.database_create'),
-                                                                'update' => trans('server/users.permissions.database_update'),
-                                                                'delete' => trans('server/users.permissions.database_delete'),
-                                                                'view_password' => trans('server/users.permissions.database_view_password'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Schedule')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.schedule_desc'))
-                                                    ->icon('tabler-clock')
-                                                    ->schema([
-                                                        CheckboxList::make('schedule')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                                'create' => 'Create',
-                                                                'update' => 'Update',
-                                                                'delete' => 'Delete',
-                                                            ])
-                                                            ->descriptions([
-                                                                'read' => trans('server/users.permissions.schedule_read'),
-                                                                'create' => trans('server/users.permissions.schedule_create'),
-                                                                'update' => trans('server/users.permissions.schedule_update'),
-                                                                'delete' => trans('server/users.permissions.schedule_delete'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Settings')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.settings_desc'))
-                                                    ->icon('tabler-settings')
-                                                    ->schema([
-                                                        CheckboxList::make('settings')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'rename' => 'Rename',
-                                                                'reinstall' => 'Reinstall',
-                                                            ])
-                                                            ->descriptions([
-                                                                'rename' => trans('server/users.permissions.setting_rename'),
-                                                                'reinstall' => trans('server/users.permissions.setting_reinstall'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                        Tab::make('Activity')
-                                            ->schema([
-                                                Section::make()
-                                                    ->description(trans('server/users.permissions.activity_desc'))
-                                                    ->icon('tabler-stack')
-                                                    ->schema([
-                                                        CheckboxList::make('activity')
-                                                            ->bulkToggleable()
-                                                            ->label('')
-                                                            ->columns(2)
-                                                            ->options([
-                                                                'read' => 'Read',
-                                                            ])
-                                                            ->descriptions([
-                                                                'read' => trans('server/users.permissions.activity_read'),
-                                                            ]),
-                                                    ]),
-                                            ]),
-                                    ]),
+                                    ->schema($tabs),
                             ]),
-                    ]),
+                    ])
+                    ->mutateRecordDataUsing(function ($data, User $user) use ($server) {
+                        $permissionsArray = $server->subusers->where('user_id', $user->id)->first()->permissions;
+
+                        $transformedPermissions = [];
+
+                        foreach ($permissionsArray as $permission) {
+                            [$group, $action] = explode('.', $permission, 2);
+                            $transformedPermissions[$group][] = $action;
+                        }
+
+                        foreach ($transformedPermissions as $key => $value) {
+                            $data[$key] = $value;
+                        }
+
+                        return $data;
+                    }),
             ]);
     }
 
