@@ -4,6 +4,7 @@ namespace App\Filament\Server\Widgets;
 
 use App\Models\Server;
 use Carbon\Carbon;
+use Filament\Facades\Filament;
 use Filament\Support\RawJs;
 use Filament\Widgets\ChartWidget;
 use Illuminate\Support\Number;
@@ -16,9 +17,19 @@ class ServerMemoryChart extends ChartWidget
 
     public ?Server $server = null;
 
+    public static function canView(): bool
+    {
+        /** @var Server $server */
+        $server = Filament::getTenant();
+
+        return !$server->isInConflictState() && !$server->retrieveStatus()->isOffline();
+    }
+
     protected function getData(): array
     {
-        $memUsed = collect(cache()->get("servers.{$this->server->id}.memory_bytes"))->slice(-10)
+        $period = auth()->user()->getCustomization()['console_graph_period'] ?? 30;
+        $memUsed = collect(cache()->get("servers.{$this->server->id}.memory_bytes"))
+            ->slice(-$period)
             ->map(fn ($value, $key) => [
                 'memory' => Number::format(config('panel.use_binary_prefix') ? $value / 1024 / 1024 / 1024 : $value / 1000 / 1000 / 1000, maxPrecision: 2),
                 'timestamp' => Carbon::createFromTimestamp($key, auth()->user()->timezone ?? 'UTC')->format('H:i:s'),
