@@ -62,7 +62,7 @@ class NetworkAllocationController extends ClientApiController
         if ($original !== $allocation->notes) {
             Activity::event('server:allocation.notes')
                 ->subject($allocation)
-                ->property(['allocation' => $allocation->toString(), 'old' => $original, 'new' => $allocation->notes])
+                ->property(['allocation' => $allocation->address, 'old' => $original, 'new' => $allocation->notes])
                 ->log();
         }
 
@@ -87,7 +87,7 @@ class NetworkAllocationController extends ClientApiController
 
         Activity::event('server:allocation.primary')
             ->subject($allocation)
-            ->property('allocation', $allocation->toString())
+            ->property('allocation', $allocation->address)
             ->log();
 
         return $this->fractal->item($allocation)
@@ -114,7 +114,7 @@ class NetworkAllocationController extends ClientApiController
 
         Activity::event('server:allocation.create')
             ->subject($allocation)
-            ->property('allocation', $allocation->toString())
+            ->property('allocation', $allocation->address)
             ->log();
 
         return $this->fractal->item($allocation)
@@ -148,7 +148,7 @@ class NetworkAllocationController extends ClientApiController
 
         Activity::event('server:allocation.delete')
             ->subject($allocation)
-            ->property('allocation', $allocation->toString())
+            ->property('allocation', $allocation->address)
             ->log();
 
         return new JsonResponse([], JsonResponse::HTTP_NO_CONTENT);
