@@ -2,22 +2,14 @@
 
 namespace App\Listeners\Auth;
 
-use Illuminate\Http\Request;
 use App\Facades\Activity;
 use Illuminate\Auth\Events\PasswordReset;
 
 class PasswordResetListener
 {
-    protected Request $request;
-
-    public function __construct(Request $request)
-    {
-        $this->request = $request;
-    }
-
     public function handle(PasswordReset $event): void
     {
-        Activity::event('event:password-reset')
+        Activity::event('auth:password-reset')
             ->withRequestMetadata()
             ->subject($event->user)
             ->log();
