@@ -12,6 +12,9 @@ use Illuminate\Http\Client\ConnectionException;
 use Sushi\Sushi;
 
 /**
+ * \App\Models\File.
+ *
+ * @property int $id
  * @property string $name
  * @property Carbon $created_at
  * @property Carbon $modified_at
@@ -26,12 +29,6 @@ use Sushi\Sushi;
 class File extends Model
 {
     use Sushi;
-
-    protected $primaryKey = 'name';
-
-    public $incrementing = false;
-
-    protected $keyType = 'string';
 
     protected int $sushiInsertChunkSize = 100;
 
@@ -153,16 +150,10 @@ class File extends Model
         try {
             $fileRepository = (new DaemonFileRepository())->setServer(self::$server);
 
-            $contents = [];
-
-            try {
-                if (!is_null(self::$searchTerm)) {
-                    $contents = cache()->remember('file_search_' . self::$path . '_' . self::$searchTerm, now()->addMinute(), fn () => $fileRepository->search(self::$searchTerm, self::$path));
-                } else {
-                    $contents = $fileRepository->getDirectory(self::$path ?? '/');
-                }
-            } catch (ConnectionException $exception) {
-                report($exception);
+            if (!is_null(self::$searchTerm)) {
+                $contents = cache()->remember('file_search_' . self::$path . '_' . self::$searchTerm, now()->addMinute(), fn () => $fileRepository->search(self::$searchTerm, self::$path));
+            } else {
+                $contents = $fileRepository->getDirectory(self::$path ?? '/');
             }
 
             if (isset($contents['error'])) {
@@ -199,8 +190,12 @@ class File extends Model
                 $message = $message->after('cURL error 7: ')->before(' after ');
             }
 
+            if ($exception instanceof ConnectionException) {
+                $message = str('Node connection failed');
+            }
+
             AlertBanner::make()
-                ->title('Could not load files')
+                ->title('Could not load files!')
                 ->body($message->toString())
                 ->danger()
                 ->send();
