@@ -4,12 +4,13 @@ namespace App\Models;
 
 use App\Contracts\Validatable;
 use App\Traits\HasValidation;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 
 class Permission extends Model implements Validatable
 {
-    use HasValidation;
+    use HasFactory, HasValidation;
 
     /**
      * The resource name for this model when it is transformed into an
@@ -38,7 +39,7 @@ class Permission extends Model implements Validatable
 
     public const ACTION_DATABASE_DELETE = 'database.delete';
 
-    public const ACTION_DATABASE_VIEW_PASSWORD = 'database.view_password';
+    public const ACTION_DATABASE_VIEW_PASSWORD = 'database.view-password';
 
     public const ACTION_SCHEDULE_READ = 'schedule.read';
 
@@ -96,6 +97,8 @@ class Permission extends Model implements Validatable
 
     public const ACTION_SETTINGS_RENAME = 'settings.rename';
 
+    public const ACTION_SETTINGS_DESCRIPTION = 'settings.description';
+
     public const ACTION_SETTINGS_REINSTALL = 'settings.reinstall';
 
     public const ACTION_ACTIVITY_READ = 'activity.read';
@@ -113,127 +116,6 @@ class Permission extends Model implements Validatable
         'permission' => ['required', 'string'],
     ];
 
-    /**
-     * All the permissions available on the system. You should use self::permissions()
-     * to retrieve them, and not directly access this array as it is subject to change.
-     *
-     * @see Permission::permissions()
-     *
-     * @var array<array-key, array{
-     *     description: string,
-     *     keys: array<array-key, string>,
-     * }>
-     */
-    protected static array $permissions = [
-        'websocket' => [
-            'description' => 'Allows the user to connect to the server websocket, giving them access to view console output and realtime server stats.',
-            'keys' => [
-                'connect' => 'Allows a user to connect to the websocket instance for a server to stream the console.',
-            ],
-        ],
-
-        'control' => [
-            'description' => 'Permissions that control a user\'s ability to control the power state of a server, or send commands.',
-            'keys' => [
-                'console' => 'Allows a user to send commands to the server instance via the console.',
-                'start' => 'Allows a user to start the server if it is stopped.',
-                'stop' => 'Allows a user to stop a server if it is running.',
-                'restart' => 'Allows a user to perform a server restart. This allows them to start the server if it is offline, but not put the server in a completely stopped state.',
-            ],
-        ],
-
-        'user' => [
-            'description' => 'Permissions that allow a user to manage other subusers on a server. They will never be able to edit their own account, or assign permissions they do not have themselves.',
-            'keys' => [
-                'create' => 'Allows a user to create new subusers for the server.',
-                'read' => 'Allows the user to view subusers and their permissions for the server.',
-                'update' => 'Allows a user to modify other subusers.',
-                'delete' => 'Allows a user to delete a subuser from the server.',
-            ],
-        ],
-
-        'file' => [
-            'description' => 'Permissions that control a user\'s ability to modify the filesystem for this server.',
-            'keys' => [
-                'create' => 'Allows a user to create additional files and folders via the Panel or direct upload.',
-                'read' => 'Allows a user to view the contents of a directory, but not view the contents of or download files.',
-                'read-content' => 'Allows a user to view the contents of a given file. This will also allow the user to download files.',
-                'update' => 'Allows a user to update the contents of an existing file or directory.',
-                'delete' => 'Allows a user to delete files or directories.',
-                'archive' => 'Allows a user to archive the contents of a directory as well as decompress existing archives on the system.',
-                'sftp' => 'Allows a user to connect to SFTP and manage server files using the other assigned file permissions.',
-            ],
-        ],
-
-        'backup' => [
-            'description' => 'Permissions that control a user\'s ability to generate and manage server backups.',
-            'keys' => [
-                'create' => 'Allows a user to create new backups for this server.',
-                'read' => 'Allows a user to view all backups that exist for this server.',
-                'delete' => 'Allows a user to remove backups from the system.',
-                'download' => 'Allows a user to download a backup for the server. Danger: this allows a user to access all files for the server in the backup.',
-                'restore' => 'Allows a user to restore a backup for the server. Danger: this allows the user to delete all the server files in the process.',
-            ],
-        ],
-
-        // Controls permissions for editing or viewing a server's allocations.
-        'allocation' => [
-            'description' => 'Permissions that control a user\'s ability to modify the port allocations for this server.',
-            'keys' => [
-                'read' => 'Allows a user to view all allocations currently assigned to this server. Users with any level of access to this server can always view the primary allocation.',
-                'create' => 'Allows a user to assign additional allocations to the server.',
-                'update' => 'Allows a user to change the primary server allocation and attach notes to each allocation.',
-                'delete' => 'Allows a user to delete an allocation from the server.',
-            ],
-        ],
-
-        // Controls permissions for editing or viewing a server's startup parameters.
-        'startup' => [
-            'description' => 'Permissions that control a user\'s ability to view this server\'s startup parameters.',
-            'keys' => [
-                'read' => 'Allows a user to view the startup variables for a server.',
-                'update' => 'Allows a user to modify the startup variables for the server.',
-                'docker-image' => 'Allows a user to modify the Docker image used when running the server.',
-            ],
-        ],
-
-        'database' => [
-            'description' => 'Permissions that control a user\'s access to the database management for this server.',
-            'keys' => [
-                'create' => 'Allows a user to create a new database for this server.',
-                'read' => 'Allows a user to view the database associated with this server.',
-                'update' => 'Allows a user to rotate the password on a database instance. If the user does not have the view_password permission they will not see the updated password.',
-                'delete' => 'Allows a user to remove a database instance from this server.',
-                'view_password' => 'Allows a user to view the password associated with a database instance for this server.',
-            ],
-        ],
-
-        'schedule' => [
-            'description' => 'Permissions that control a user\'s access to the schedule management for this server.',
-            'keys' => [
-                'create' => 'Allows a user to create new schedules for this server.', // task.create-schedule
-                'read' => 'Allows a user to view schedules and the tasks associated with them for this server.', // task.view-schedule, task.list-schedules
-                'update' => 'Allows a user to update schedules and schedule tasks for this server.', // task.edit-schedule, task.queue-schedule, task.toggle-schedule
-                'delete' => 'Allows a user to delete schedules for this server.', // task.delete-schedule
-            ],
-        ],
-
-        'settings' => [
-            'description' => 'Permissions that control a user\'s access to the settings for this server.',
-            'keys' => [
-                'rename' => 'Allows a user to rename this server and change the description of it.',
-                'reinstall' => 'Allows a user to trigger a reinstall of this server.',
-            ],
-        ],
-
-        'activity' => [
-            'description' => 'Permissions that control a user\'s access to the server activity logs.',
-            'keys' => [
-                'read' => 'Allows a user to view the activity logs for the server.',
-            ],
-        ],
-    ];
-
     protected function casts(): array
     {
         return [
@@ -242,10 +124,91 @@ class Permission extends Model implements Validatable
     }
 
     /**
+     * All the permissions available on the system.
+     *
+     * @return array<int, array{
+     *      name: string,
+     *      icon: string,
+     *      permissions: string[]
+     *  }>
+     */
+    public static function permissionData(): array
+    {
+        return [
+            [
+                'name' => 'control',
+                'icon' => 'tabler-terminal-2',
+                'permissions' => ['console', 'start', 'stop', 'restart'],
+            ],
+            [
+                'name' => 'user',
+                'icon' => 'tabler-users',
+                'permissions' => ['read', 'create', 'update', 'delete'],
+            ],
+            [
+                'name' => 'file',
+                'icon' => 'tabler-files',
+                'permissions' => ['read', 'read-content', 'create', 'update', 'delete', 'archive', 'sftp'],
+            ],
+            [
+                'name' => 'backup',
+                'icon' => 'tabler-file-zip',
+                'permissions' => ['read', 'create', 'delete', 'download', 'restore'],
+            ],
+            [
+                'name' => 'allocation',
+                'icon' => 'tabler-network',
+                'permissions' => ['read', 'create', 'update', 'delete'],
+            ],
+            [
+                'name' => 'startup',
+                'icon' => 'tabler-player-play',
+                'permissions' => ['read', 'update', 'docker-image'],
+            ],
+            [
+                'name' => 'database',
+                'icon' => 'tabler-database',
+                'permissions' => ['read', 'create', 'update', 'delete', 'view-password'],
+            ],
+            [
+                'name' => 'schedule',
+                'icon' => 'tabler-clock',
+                'permissions' => ['read', 'create', 'update', 'delete'],
+            ],
+            [
+                'name' => 'settings',
+                'icon' => 'tabler-settings',
+                'permissions' => ['rename', 'description', 'reinstall'],
+            ],
+            [
+                'name' => 'activity',
+                'icon' => 'tabler-stack',
+                'permissions' => ['read'],
+            ],
+        ];
+    }
+
+    /**
      * Returns all the permissions available on the system for a user to have when controlling a server.
      */
     public static function permissions(): Collection
     {
-        return Collection::make(self::$permissions);
+        $permissions = [
+            'websocket' => [
+                'description' => 'Allows the user to connect to the server websocket, giving them access to view console output and realtime server stats.',
+                'keys' => [
+                    'connect' => 'Allows a user to connect to the websocket instance for a server to stream the console.',
+                ],
+            ],
+        ];
+
+        foreach (static::permissionData() as $data) {
+            $permissions[$data['name']] = [
+                'description' => trans('server/users.permissions.' . $data['name'] . '_desc'),
+                'keys' => collect($data['permissions'])->mapWithKeys(fn ($key) => [$key => trans('server/users.permissions.' . $data['name'] . '_' . str($key)->replace('-', '_'))])->toArray(),
+            ];
+        }
+
+        return collect($permissions);
     }
 }
