@@ -265,8 +265,13 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
      */
     public function accessibleServers(): Builder
     {
-        if ($this->canned('viewList server')) {
-            return Server::query();
+        if ($this->canned('viewAny', Server::class)) {
+            return Server::select('servers.*')
+                ->leftJoin('subusers', 'subusers.server_id', '=', 'servers.id')
+                ->where(function (Builder $builder) {
+                    $builder->where('servers.owner_id', $this->id)->orWhere('subusers.user_id', $this->id)->orWhereIn('servers.node_id', $this->accessibleNodes()->pluck('id'));
+                })
+                ->distinct('servers.id');
         }
 
         return $this->directAccessibleServers();
@@ -278,12 +283,27 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
      */
     public function directAccessibleServers(): Builder
     {
-        return Server::query()
-            ->select('servers.*')
+        return Server::select('servers.*')
             ->leftJoin('subusers', 'subusers.server_id', '=', 'servers.id')
             ->where(function (Builder $builder) {
                 $builder->where('servers.owner_id', $this->id)->orWhere('subusers.user_id', $this->id);
             });
+    }
+
+    public function accessibleNodes(): Builder
+    {
+        // Root admins can access all nodes
+        if ($this->isRootAdmin()) {
+            return Node::query();
+        }
+
+        // Check if there are no restrictions from any role
+        $roleIds = $this->roles()->pluck('id');
+        if (!NodeRole::whereIn('role_id', $roleIds)->exists()) {
+            return Node::query();
+        }
+
+        return Node::whereHas('roles', fn (Builder $builder) => $builder->whereIn('roles.id', $roleIds));
     }
 
     public function subusers(): HasMany
@@ -298,12 +318,12 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
 
     protected function checkPermission(Server $server, string $permission = ''): bool
     {
-        if ($this->canned('update server', $server) || $server->owner_id === $this->id) {
+        if ($this->canned('update', $server) || $server->owner_id === $this->id) {
             return true;
         }
 
         // If the user only has "view" permissions allow viewing the console
-        if ($permission === Permission::ACTION_WEBSOCKET_CONNECT && $this->canned('view server', $server)) {
+        if ($permission === Permission::ACTION_WEBSOCKET_CONNECT && $this->canned('view', $server)) {
             return true;
         }
 
@@ -390,13 +410,24 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
         return $provider?->get($this);
     }
 
-    public function canTarget(Model $user): bool
+    public function canTarget(Model $model): bool
     {
+        // Root admins can target everyone and everything
         if ($this->isRootAdmin()) {
             return true;
         }
 
-        return $user instanceof User && !$user->isRootAdmin();
+        // Make sure normal admins can't target root admins
+        if ($model instanceof User) {
+            return !$model->isRootAdmin();
+        }
+
+        // Make sure the user can only target accessible nodes
+        if ($model instanceof Node) {
+            return $this->accessibleNodes()->where('id', $model->id)->exists();
+        }
+
+        return false;
     }
 
     public function getTenants(Panel $panel): array|Collection
@@ -407,7 +438,7 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
     public function canAccessTenant(Model $tenant): bool
     {
         if ($tenant instanceof Server) {
-            if ($this->canned('view server', $tenant) || $tenant->owner_id === $this->id) {
+            if ($this->canned('view', $tenant) || $tenant->owner_id === $this->id) {
                 return true;
             }
 
