@@ -15,8 +15,9 @@ return [
         'startup_read' => 'Allows a user to view the startup variables for a server.',
         'startup_update' => 'Allows a user to modify the startup variables for the server.',
         'startup_docker_image' => 'Allows a user to modify the Docker image used when running the server.',
-        'setting_reinstall' => 'Allows a user to trigger a reinstall of this server.',
-        'setting_rename' => 'Allows a user to rename this server and change the description of it.',
+        'settings_reinstall' => 'Allows a user to trigger a reinstall of this server.',
+        'settings_rename' => 'Allows a user to rename this server.',
+        'settings_description' => 'Allows a user to change the description of this server.',
         'activity_read' => 'Allows a user to view the activity logs for the server.',
         'websocket_*' => 'Allows a user access to the websocket for this server.',
         'control_console' => 'Allows a user to send data to the server console.',
