@@ -1,11 +1,25 @@
 <x-filament::widget>
     @assets
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@xterm/xterm/css/xterm.min.css">
+    @php
+        $userFont = auth()->user()->getCustomization()['console_font'] ?? 'monospace';
+        $userFontSize = auth()->user()->getCustomization()['console_font_size'] ?? 14;
+        $userRows =  auth()->user()->getCustomization()['console_rows'] ?? 30;
+    @endphp
+    @if($userFont !== "monospace")
+        <link rel="preload" href="{{ asset("storage/fonts/{$userFont}.ttf") }}" as="font" crossorigin>
+        <style>
+            @font-face {
+                font-family: '{{ $userFont }}';
+                src: url('{{ asset("storage/fonts/{$userFont}.ttf") }}');
+            }
+        </style>
+    @endif
     <script src="https://cdn.jsdelivr.net/npm/@xterm/xterm/lib/xterm.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@xterm/addon-fit/lib/addon-fit.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@xterm/addon-web-links/lib/addon-web-links.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@xterm/addon-search/lib/addon-search.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/xterm-addon-search-bar/lib/xterm-addon-search-bar.min.js"></script>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@xterm/xterm/css/xterm.min.css">
     <link rel="stylesheet" href="{{ asset('/css/filament/server/console.css') }}">
     @endassets
 
@@ -57,14 +71,14 @@
         };
 
         let options = {
-            fontSize: {{ auth()->user()->getCustomization()['console_font_size'] ?? 14 }},
-            fontFamily: 'Comic Mono, monospace',
+            fontSize: {{ $userFontSize }},
+            fontFamily: '{{ $userFont }}, monospace',
             lineHeight: 1.2,
             disableStdin: true,
             cursorStyle: 'underline',
             cursorInactiveStyle: 'underline',
             allowTransparency: true,
-            rows: {{ auth()->user()->getCustomization()['console_rows'] ?? 30 }},
+            rows: {{ $userRows }},
             theme: theme
         };
 
